Fix: auto-escaping in entry-image.
This patch apply the guidelines for upgrading to 1.6.

Indeed, after testing this template on a fresh install of Grav, I couldn't visually see the "entry-image" without converting to raw. 

As explained through the migration guide https://learn.getgrav.org/17/advanced/grav-development/grav-16-upgrade-guide#auto-escaping 

"The transition to use auto-escaping will not be easy. During the transition all the template files should either contain both `|e` and `|raw` filters on every variable to make sure that the template file is safe to be used in both modes, or you can surround all the template code with `{% autoescape %}` Twig tags."